# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: 한국 여행 가이드 API
  description: OAuth 기반 인증, AI 채팅, Guest-Guide 매칭을 지원하는 여행 가이드 서비스 (MVP)
  version: 2.0.0
  contact:
    name: 개발팀
    email: dev@example.com
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
...
components:
  schemas:
    ApiResponse:
      type: object
      properties:
        msg:
          type: string
          description: 응답 메시지
        data:
          type: object
          description: 응답 데이터
          nullable: true

    # Auth Schemas
    UserRole:
      type: string
      enum: [USER, ADMIN, GUIDE, PENDING]
      description: 사용자 역할
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        nickname:
          type: string
        profileImageUrl:
          type: string
          format: uri
          nullable: true
        role:
          $ref: '#/components/schemas/UserRole'
    UserUpdateRequest:
      type: object
      properties:
        nickname:
          type: string
          nullable: true
        profileImageUrl:
          type: string
          format: uri
          nullable: true

    # AI Chat Domain Schemas
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
paths:
...
  /api/users/me:
    get:
      tags:
        - users
      summary: 내 정보 조회
      description: 현재 로그인된 사용자의 프로필 정보를 조회합니다.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 내 정보 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
    patch:
      tags:
        - users
      summary: 내 프로필 수정
      description: 현재 로그인된 사용자의 프로필 정보(닉네임, 프로필 이미지)를 수정합니다.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 프로필 수정 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserResponse'
    delete:
      tags:
        - users
      summary: 회원 탈퇴
      description: 현재 로그인된 사용자 계정을 영구적으로 삭제합니다.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 회원 탈퇴 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  # ===================
  # AI Chat Domain APIs
  # ===================
...